refactor: 将工具调用逻辑移动到中间件文件
